+++ arduino/lib/aop/PunchCard.h
@@ -14,6 +14,7 @@ namespace AOP {
 class PunchCard
 {
 public:
+    static constexpr const uint8_t CLEAR_STATION = 0;
     static constexpr const uint8_t CHECK_STATION = 1;
     static constexpr const uint8_t START_STATION = 10;
     static constexpr const uint8_t FINISH_STATION = 255;